Move modules
moved checked_dict and md5hash to the tools package
 and fixed references
