+++ snappy/swift_behaviors.py
@@ -11,7 +11,7 @@ from StringIO import StringIO
 from time import sleep, time
 
 from unicode_helpers import UNICODE_BLOCKS, BLOCK_NAMES
-from md5hash import get_md5_hash
+from snappy.tools.md5hash import get_md5_hash
 from os_client import ObjectStorageAPIClient
 from constants import Constants
