+++ snappy/tools/check_dict.py
